re #757: Fixed slowing down of paint effect
As more and more spheres were added, paint effect was slowing down. The problem was that each time the mouse moved all sphere actors were removed and added back.

Fixed by using a single actor for feedback with a glyph filter (that clones the mouse cursor N times).

git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2748 25a7314f-c4c8-4314-9936-a25f8480a772

+++ Segmentations/EditorEffects/qSlicerSegmentEditorAbstractEffect.cxx
@@ -791,7 +791,7 @@ void qSlicerSegmentEditorAbstractEffect::xyToRas(QPoint xy, double outputRas[3],
 }
 
 //-----------------------------------------------------------------------------
-void qSlicerSegmentEditorAbstractEffect::xyToRas(int xy[2], double outputRas[3], qMRMLSliceWidget* sliceWidget)
+void qSlicerSegmentEditorAbstractEffect::xyToRas(double xy[2], double outputRas[3], qMRMLSliceWidget* sliceWidget)
 {
   double xyz[3] = {xy[0], xy[1], 0.0};
   qSlicerSegmentEditorAbstractEffect::xyzToRas(xyz, outputRas, sliceWidget);
@@ -851,7 +851,7 @@ void qSlicerSegmentEditorAbstractEffect::xyToIjk(QPoint xy, int outputIjk[3], qM
 }
 
 //-----------------------------------------------------------------------------
-void qSlicerSegmentEditorAbstractEffect::xyToIjk(int xy[2], int outputIjk[3], qMRMLSliceWidget* sliceWidget, vtkOrientedImageData* image)
+void qSlicerSegmentEditorAbstractEffect::xyToIjk(double xy[2], int outputIjk[3], qMRMLSliceWidget* sliceWidget, vtkOrientedImageData* image)
 {
   double xyz[3] = {xy[0], xy[0], 0.0};
   qSlicerSegmentEditorAbstractEffect::xyzToIjk(xyz, outputIjk, sliceWidget, image);
